BACKPORT: UPSTREAM: scsi: ufs: qcom: Rename "hs_gear" to "phy_gear"
The "hs_gear" variable is used to cache the gear setting for the
PHY that will be used during ufs_qcom_power_up_sequence(). But it
creates ambiguity with the gear setting used by the ufshcd driver.

So let's rename it to "phy_gear" to make it explicit that this
variable caches the gear setting for the PHY.

Change-Id: Ic00005a7e53913bc220dff3c46663dc317248b04
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Link: https://lore.kernel.org/r/20230908145329.154024-2-manivannan.sadhasivam@linaro.org
Reviewed-by: Can Guo <quic_cang@quicinc.com>
Tested-by: Can Guo <quic_cang@quicinc.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Git-Commit: 5a738cfe49b2cc0896353c0b33dc5cc81316aabe
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Ram Kumar Dwivedi <quic_rdwivedi@quicinc.com>