BACKPORT: UPSTREAM: scsi: ufs: qcom: Clarify comments about the initial phy_gear
The comments that currently are within the hw_ver < 4 conditional are
misleading. They really apply to various branches of the conditionals
there and incorrectly state that the phy_gear value can increase.

Right now the logic is to:

 - Default to max supported gear for phy_gear

 - Set phy_gear to minimum value if version < 4 since those versions
   only
   support one PHY init sequence (and therefore don't need reinit)

 - Set phy_gear to the optimal value if the device version is already
   populated in the controller registers on boot

Let's move some of the comment to outside the if statement and clean up the
bit left about switching to a higher gear on reinit.
This way the comment more accurately reflects the logic.

Change-Id: I78df18131302b7696d516823084ed7cb840d3d8e
Signed-off-by: Andrew Halaney <ahalaney@redhat.com>
Link: https://lore.kernel.org/r/20240123-ufs-reinit-comments-v1-1-ff2b3532d7fe@redhat.com
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Git-Commit: 883a8b451cf4c659d5bba0becfc3780394b597c8
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Ram Kumar Dwivedi <quic_rdwivedi@quicinc.com>