FROMLIST: pstore/ram: Add ramoops information notifier support
Client like minidump is interested in knowing ramoops individual zone
addresses and their size so that it could register them with its table.

Let's introduce a info notifier in ramoops which gets called when
ramoops driver probes successfully and it passes the ramoops region
information to the passed callback by the client and If the call for
ramoops ready register comes after ramoops probe than call the callback
directly.

Change-Id: I230cf29d87aacc4ee4037d8b8eff6b1afd3965b0
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Link: https://lore.kernel.org/lkml/20240131110837.14218-10-quic_mojha@quicinc.com/
Signed-off-by: Mukesh Ojha <mukesh.ojha@oss.qualcomm.com>