PENDING: ptp: qcom: Add support for Time Stamp Counter(TSC) driver
TSC module is a timestamp generator, which runs a 64 bit
counter for system timekeeping.

ETU is event timestamp module, which captures the timestamp from TSC
and global counter bus for a given event. The event is selected from
a set of GPIOs and SW triggered events. The design is parameterized
to detect multiple events, using different “slices”.
One slice is used for every event.

Change-Id: Ibbf89ca2797bd4fdc1858ac408248c294321c2c7
Signed-off-by: Taniya Das <quic_tdas@quicinc.com>
Signed-off-by: Satya Priya Kakitapalli <quic_skakitap@quicinc.com>