BACKPORT: FROMLIST: dt-bindings: remoteproc: qcom,pas-common: Introduce iommus and qcom,devmem property
Qualcomm’s PAS implementation for remote processors only supports a
single stage of IOMMU translation and is presently managed by the
Qualcomm EL2 hypervisor (QHEE) if it is present. In the absence of QHEE,
such as with a KVM hypervisor, IOMMU translations need to be set up by
the KVM host. Remoteproc needs carveout memory region and its resource
(device memory) permissions to be set before it comes up, and this
information is presently available statically with QHEE.

In the absence of QHEE, the boot firmware needs to overlay this
information based on SoCs running with either QHEE or a KVM hypervisor
(CPUs booted in EL2).

The qcom,devmem property provides IOMMU devmem translation information
intended for non-QHEE based systems.

Change-Id: I3ffcabb379ba6acf6759cabff96eee80fc388155
Signed-off-by: Shiraz Hashim <quic_shashim@quicinc.com>
Co-developed-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Link: https://lore.kernel.org/lkml/20241004212359.2263502-2-quic_mojha@quicinc.com/
[mukesh.ojha: s/Co-Developed-by/Co-developed-by to make checkpatch.pl]
Signed-off-by: Mukesh Ojha <mukesh.ojha@oss.qualcomm.com>