QCLINUX: usb: dwc3: qcom: Add shutdown callback for qcom wrapper
This patch adds a shutdown callback to USB DWC core driver to
ensure that it is properly shutdown in reboot/shutdown path. If
the hardware is still accessing memory after SMMU translation is
disabled as part of SMMU shutdown callback in system reboot or
shutdown path, then IOVAs(I/O virtual address) which it was using
will go on the bus as the physical addresses which might result
in unknown crashes (NoC/interconnect errors). Moreover add this
callback only for flattened use cases where the issue was seen so
far.

Change-Id: Iad45ff612340cdfbc09af90e1b3331406d9f70b8
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>