FROMGIT: arm64: defconfig: enable clock controller, interconnect and pinctrl for QCS615
Enable clock controller, interconnect and pinctrl for Qualcomm QCS615
platform to boot to UART console.

The serial engine depends on GCC, interconnect, and pinctrl. It is
necessary to build them as built-in modules because the debug console
must be registered before userspace is launched. The primary reason for
this is that, for example, systemd opens /dev/console at launch (i.e.,
when the init process starts). Therefore, if we register the console after
this, we will not receive console output from systemd.

Change-Id: I94949899c16962b6aa7b17ff434ca918ff3a8b48
Signed-off-by: Lijuan Gao <quic_lijuang@quicinc.com>
Reviewed-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Link: https://lore.kernel.org/r/20241104-add_initial_support_for_qcs615-v5-5-9dde8d7b80b0@quicinc.com
Signed-off-by: Bjorn Andersson <andersson@kernel.org>
Git-commit: 9eec6ce36b5dc981327e9f58025d012e524687b4
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git