FROMLIST: drm/msm/dp: make bridge helpers use dp_display to allow re-use
dp_bridge helpers take drm_bridge as an input and extract the
dp_display object to be used in the dp_display module. Rather than
doing it in a roundabout way, directly pass the dp_display object
to these helpers so that the MST bridge can also re-use the same
helpers.

Change-Id: Id2862dcd05dad8b0b9d00d5e955a6c4b37dd8736
Link: https://lore.kernel.org/all/20241205-dp_mst-v1-12-f8618d42a99a@quicinc.com/
Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Signed-off-by: Yongxing Mou <quic_yongmou@quicinc.com>