FROMLIST: drm/msm/dp: disable the opp table request even for dp_ctrl_off_link()
dp_ctrl_off_link() was created to handle a case where we
received a cable connection and then get a cable disconnect
without the corresponding dp_display_enable(). For such
cases the px clock will be off but the link clock will still
be on. dp_ctrl_off_link() handles this case by turning off
the link clock only.

However, the vote removal to the opp table for this case was
missed.Remove the opp table vote in dp_ctrl_off_link().

Change-Id: I3ebf259f5b6b398dc4e2664e56e95804048427b0
Link: https://lore.kernel.org/all/20241205-dp_mst-v1-2-f8618d42a99a@quicinc.com/
Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Signed-off-by: Yongxing Mou <quic_yongmou@quicinc.com>