FROMLIST: drm/msm/dp: introduce stream_id for each DP panel
With MST, each DP controller can handle multiple streams.
There shall be one dp_panel for each stream but the dp_display
object shall be shared among them. To represent this abstraction,
create a stream_id for each DP panel which shall be set by the
MST stream. For SST, default this to stream 0.

Use the stream ID to control the px clock of that respective
stream by extending the clock handles and state tracking of the
DP px clock to an array of max supported streams.

Change-Id: I8caa77c254ca4cc88e6778996de1edcd0b6fd95e
Link: https://lore.kernel.org/all/20241205-dp_mst-v1-14-f8618d42a99a@quicinc.com/
Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Signed-off-by: Yongxing Mou <quic_yongmou@quicinc.com>