FROMLIST: drm/msm/dp: add support to send ACT packets for MST
Whenever virtual channel slot allocation changes, the DP
source must send the action control trigger sequence to notify
the sink about the same. This would be applicable during the
start and stop of the stream. Add the infrastructure
to be able to send ACT packets for the DP controller when
operating in MST mode.

Change-Id: I539ec41f300acee063bf065ad56db105bf94afe1
Link: https://lore.kernel.org/all/20241205-dp_mst-v1-18-f8618d42a99a@quicinc.com/
Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Signed-off-by: Yongxing Mou <quic_yongmou@quicinc.com>