FROMLIST: drm/msm/dp: add support to re-use and clear the panel edid
During certain cases, the dp mst connector's panel edid can be
re-used such as getting multiple get_modes() without irq_hpd and
should be cleared in cases when the connector is destroyed or when
irq hpd is received. Add support to handle these cases for the
mst_connector's panel edid.

Change-Id: I0ac0bc4262771cc91f6c8ac9883d67634a605154
Link: https://lore.kernel.org/all/20241205-dp_mst-v1-34-f8618d42a99a@quicinc.com/
Signed-off-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Signed-off-by: Yongxing Mou <quic_yongmou@quicinc.com>