PENDING: arm64: dts: qcom: Disable USB U1/U2 entry for QCS9100/QCS8300
Enabling U1 and U2 power-saving states can lead to stability and
performance issues, particularly for latency-sensitive or high-
throughput applications. These low-power link states are intended
to reduce power consumption by allowing the device to enter partial
low-power modes during idle periods. However, they can sometimes
result in unexpected behavior. Over the years, some of the issues
seen are as follows:

On QCS8300/QCS9100, it is observed that when Link enters U2, when
the cable is disconnected and reconnected to host PC in HS, there
is no link status change interrupt seen and the plug-in in HS does
not show up a bus reset and enumeration failure happens.

Change-Id: I3ae9e6ba27c746e9b4b8725fbc98879b485c6d0e
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>