PENDING: phy: qcom: qmp: Add default Type-C reverse orientation flag
Add support for a default Type-C reverse orientation flag.
Based on this flag, "default-typec-orientation-reverse," a reverse Type-C
orientation can be set if the reverse lane is fixed for the USB port.

In the current implementation of the QMP Combo driver, the initial Type-C
orientation is set to Lane_A, and the driver relies on calls from glink to
inform of orientation changes via the Type-C framework. However, on some
platforms, Lane_B is routed to the USB port, and Lane_A is routed to the
display. If the orientation detection GPIO, which is usually routed to the
SoC, is not present, there would be no orientation switch calls to the QMP
driver. Since the default orientation is set to Lane_A, USB SuperSpeed
would be non-functional.

To support such platforms, this patch adds support to select the default
reverse Type-C orientation during the phy_init probe.

Change-Id: If876f5f2c5957799d28df60f08627e7c8d6c3fa8
Signed-off-by: Akash Kumar <quic_akakum@quicinc.com>