BACKPORT: PCI: dwc: Skip waiting for link up if vendor drivers can detect Link up event
If the vendor drivers can detect the Link up event using mechanisms
such as Link up IRQ and if the driver can enumerate downstream devices
instead of waiting here, then waiting for Link up during probe is not
needed here, which optimizes the boot time.

So skip waiting for link to be up if the driver supports 'use_linkup_irq'.

Change-Id: I1f85560d04a180a77a945dd9dfaac94d5f76cb89
Link: https://lore.kernel.org/all/20241123-remove_wait2-v5-1-b5f9e6b794c2@quicinc.com/
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Reviewed-by: Niklas Cassel <cassel@kernel.org>
Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>