BACKPORT: PCI: qcom: Set use_linkup_irq if global IRQ handler is present
In cases where a global IRQ handler is present to manage link up
interrupts, it may not be necessary to wait for the link to be up
during PCI initialization which optimizes the bootup time.

So, set use_linkup_irq flag if global IRQ is present and in order to set
the use_linkup_irq flag before calling dw_pcie_host_init() call, which
waits for link to be up, move platform_get_irq_byname_optional() API
above dw_pcie_host_init().

Change-Id: Id4d0af4cce6487ae0661c968f25d40c55faf44a6
Link: https://lore.kernel.org/all/20241123-remove_wait2-v5-2-b5f9e6b794c2@quicinc.com/
Reviewed-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Reviewed-by: Niklas Cassel <cassel@kernel.org>
Signed-off-by: Krishna chaitanya chundru <quic_krichai@quicinc.com>