QCLINUX: arm64: dts: qcom: Update the size for swao-etf node
The swao-etf node in memory dump table should have same size with
default size of ETB buffer. The default size for ETB buffer is
0x10000. Therefore, update the size for swao-etf node for QCS8300
and SA8775p.

Remove tmc-etf and etf-reg nodes from memory dump table for QCS8300
and SA8775p, as this old device is no longer in use.

Change-Id: I6daed4bdaaf11473f777bc74e2d6ee5c488d9e30
Signed-off-by: Jie Gan <quic_jiegan@quicinc.com>