PENDING: drm/bridge: anx7625: update bridge_ops and sink detect logic
In the anx7625_link_bridge function, we verify if a specific device
is a panel bridge. If it is not, we add DRM_BRIDGE_OP_HPD or
DRM_BRIDGE_OP_DETECT to the bridge operations. However, on port 1
of the anx7625 bridge, any device added is always treated as a
panel bridge, which prevents our connector_detect function from
being called. To address this, instead of merely checking whether
it is a panel bridge. we should verify the type of the panel
bridge—whether it is a DisplayPort or eDP panel. If the panel
bridge is a DisplayPort type, then we should add
DRM_BRIDGE_OP_HPD or DRM_BRIDGE_OP_DETECT to the bridge
operations.

In the anx7625_sink_detect function, it checks if the device
is a panel bridge and always sends a "connected" statusto the
connector. When adding the DP port on port 1 of the anx7625,
it incorrectly treats it as a panel bridge and sends an
always "connected" status. Instead of checking the status
on the panel bridge, it's better to check the hpd_status
for connectors like DisplayPort. This way, it verifies the
hpd_status variable before sending the status to the
connector. For other connectors like eDP panels, the
hpd_status variable will always be greater than 0.

Change-Id: I26fdce1847f0785d32f198c11e927cf529912189
Signed-off-by: Ayushi Makhija <quic_amakhija@quicinc.com>