FROMLIST: dt-bindings: net: qcom,ethqos: Drop fallback compatible for qcom,qcs615-ethqos
The core version of EMAC on qcs615 has minor differences compared to that
on sm8150. During the bring-up routine, the loopback bit needs to be set,
and the Power-On Reset (POR) status of the registers isn't entirely
consistent with sm8150 either.
Therefore, it should be treated as a separate entity rather than a
fallback option.

Fixes: 32535b9410b8 ("dt-bindings: net: qcom,ethqos: add description for qcs615")
Link: https://lore.kernel.org/all/20241224-schema-v2-0-000ea9044c49@quicinc.com/
Change-Id: Ia83fb6058b68eec4413e980484ad4c35dc326c55
Signed-off-by: Yijie Yang <quic_yijiyang@quicinc.com>