FROMLIST: misc: fastrpc: Fix race condition in MUNMAP
fastrpc_req_munmap involves two steps to unmap memory, first to locates
a matching fastrpc buf in the list and second is to send request to DSP
to unmap. There is a potential race condition between two operations
which can lead to user-after-free scenario.
Lock unmap request to avoid use-after-free.

Change-Id: Ibbf2a6375387087cd1386cda064f28f5d147a0d0
Link: https://lore.kernel.org/linux-arm-msm/20250108044802.3829060-1-quic_lxu5@quicinc.com/
Signed-off-by: Ling Xu <quic_lxu5@quicinc.com>