PENDING: net: stmmac: Enable PTP Packet id based timestamping
Enable PTP support to add Packet id based timestamping. The
configuration of ptp packets are done through context descriptors
and configuring the packet id and read in a fifo fashion from the
dma rings. Submitting the PTP timestamped packets to the DMA
and retrieving the timestamp packets after interrupt completion and
passing it to network stack.

Change-Id: Ifa58844e3d6d3572569504a99418f4cbb1c09db3
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>