PENDING: drm/msm/dp: add catalog support for 4 streams MST
Add support for the pixel and link registers related to 4 streams MST in
the catalog. The pixel2 and pixel3 clock registers have the same
definitions. The MST2/MST3 link clk for pixel2 and pixel3 have different
offset and definition compare to main link clk. Add the corresponding
macro definitions in dp_reg.h.

Change-Id: Ia07843e36e140657e54188456266c7f2d68f1948
Signed-off-by: Yongxing Mou <quic_yongmou@quicinc.com>