FROMLIST: arm64: dts: qcom: qcs8300-ride: Add firmware-name in BT node
The qcs8300-ride platform uses the QCA6698 Bluetooth chip. While the
QCA6698 shares the same IP core as the WCN6855, it has different RF
components and RAM sizes, requiring new firmware files. Use the
firmware-name property to specify the NVM and rampatch firmware to load.

Note: Only the firmware-name is being added, as the rest of the
changes have already been merged.

Link: https://lore.kernel.org/linux-arm-msm/20250211104421.1172892-2-quic_chejiang@quicinc.com/
Change-Id: I9f83f99e68ca5741ab2d3fc4595f1e30b31c8931
Signed-off-by: Cheng Jiang <quic_chejiang@quicinc.com>