PENDING: arm64: dts: qcom: qcs9075: Change phy-mode to SGMII
Current QCA8081 phy driver on QCS9075-RB8 supports only 2.5G link speed.
If a lower capability link parnter is connected, the data path breaks.
2.5 only link partners are not commonly used. This is a problem as one
expects an ethernet port to work at a variety of speeds. Force the phy
to operate in SGMII mode so that the data path will work on 100M/1G link
speeds (which are more commonly used).

Change-Id: I03bd281c4e80ff5a8482ac2d7a3ac569ac717598
Signed-off-by: Mohd Ayaan Anwar <quic_mohdayaa@quicinc.com>