UPSTREAM: arm64: dts: qcom: sc7280: Enable MDP turbo mode
The max frequency listed in the DPU opp-table is 506MHz, this is not
sufficient to drive a 4k@60 display, resulting in constant underrun.

Add the missing MDP_CLK turbo frequency of 608MHz to the opp-table to
fix this.

Change-Id: Ibb36b41f7fcbabfffb0654c1fdc749835f77abb1
Signed-off-by: Bjorn Andersson <quic_bjorande@quicinc.com>
Reviewed-by: Abhinav Kumar <quic_abhinavk@quicinc.com>
Reviewed-by: Konrad Dybcio <konrad.dybcio@linaro.org>
Acked-by: Douglas Anderson <dianders@chromium.org>
Link: https://lore.kernel.org/all/20240326-rb3gen2-dp-connector-v2-1-a9f1bc32ecaf@quicinc.com
Git-Commit: d0d6230aa9652afc2c96a83eb2d217761683cad8
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Amit Singh <quic_amitsi@quicinc.com>