BACKPORT: UPSTREAM: cpuidle: psci: Enable the hierarchical topology for s2ram on PREEMPT_RT
The hierarchical PM domain topology are currently disabled on a
PREEMPT_RT based configuration. As a first step to enable it to be
used, let's try to attach the CPU devices to their PM domains on
PREEMPT_RT. In this way the syscore ops becomes available, allowing
the PM domain topology to be managed during s2ram.

For the moment let's leave the support for CPU hotplug outside
PREEMPT_RT, as it's depending on using runtime PM. For s2ram, this
isn't a problem as all CPUs are managed via the syscore ops.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Tested-by: Raghavendra Kakarla <quic_rkakarla@quicinc.com>  # qcm6490 with PREEMPT_RT set
Acked-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Link: https://lore.kernel.org/r/20240527142557.321610-7-ulf.hansson@linaro.org
Git-commit: 4517b1c383807fadac8374f99f2361fe7eb4c0b4
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Change-Id: Ic3d2f397b59a47cdece1fb2ddb0dc51142fa86f9
Signed-off-by: Raghavendra Kakarla <quic_rkakarla@quicinc.com>