PENDING: misc: fastrpc: Remove PD status check
PD state check is added for static PD invocations. This state is
set by PDR notifications which requires PD-mapper service. If
there is any problem with PD-mapper service or PDR interface, this
check fails which results in failure of audio or sensors daemon
connection. Remove this check as PD restart feature is not
required.

Change-Id: I7fd1a704d3f2a84cfb6085b9e256dfd1b0cd380c
Signed-off-by: Ekansh Gupta <quic_ekangupt@quicinc.com>