PENDING: misc: eud: Add High-Speed Phy control for EUD operations
The Embedded USB Debugger(EUD) is a HS-USB on-chip hub to support the
debug and trace capabilities on Qualcomm devices. It sits in between
the usb port and the usb controller. Being a HS USB hub, it relies on
HS Phy for its functionality. Add HS phy support in the eud driver
and control the phy during eud enable/disable operations.

Change-Id: Ica4377c63876206da6d5942d4f6ad718c0729980
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>
Signed-off-by: Krishna Kurapati <krishna.kurapati@oss.qualcomm.com>