PENDING: net: stmmac: Use generic PCS field and move pcs_pre_init before Serdes powerup
As PCS models are hardware-specific, it is more
appropriate to use the generic PCS field in the
mac_device_info struct, remove the usage of
qcom_xpcs everywhere and use the generic phylink_pcs
instead.
When STMMAC_FLAG_SERDES_UP_AFTER_PHY_LINKUP dev flag is
not set we expect Serdes programming to be done within
ndo_open callback and that requires Rx clock to be present.
So make sure enabling of Xpcs Rx loopback clock is done
before we start programming Serdes block not after that.

Change-Id: I891b31e7e591ecb6408774edc5f69949fd9c858e
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>