FROMLIST: firmware: qcom_scm: Check for waitq state in wait_for_completion
Modify wait_for_completion() to check if task is in idle state using
wait_for_completion_state().

This allows for detecting when waitq contexts are in idle state and
propagates it to __scm_smc_do(), which is beneficial when task is idle
and waiting for a kick to accept new requests.


Link: https://lore.kernel.org/lkml/20250425-multi_waitq_scm-v6-3-cba8ca5a6d03@oss.qualcomm.com/
Change-Id: Ic30549ab6baf6f077114d96f50e4a0ce43bca02d
Signed-off-by: Unnathi Chalicheemala <unnathi.chalicheemala@oss.qualcomm.com>
Signed-off-by: Yuvaraj Ranganathan <quic_yrangana@quicinc.com>
Signed-off-by: Ninad Naik <quic_ninanaik@quicinc.com>