BACKPORT: arm64: dts: qcom: qcs615: fix a crash issue caused by infinite loop for Coresight
An infinite loop has been created by the Coresight devices. When only a
source device is enabled, the coresight_find_activated_sysfs_sink function
is recursively invoked in an attempt to locate an active sink device,
ultimately leading to a stack overflow and system crash. Therefore, disable
the replicator1 to break the infinite loop and prevent a potential stack
overflow.

Fixes: bf469630552a ("arm64: dts: qcom: qcs615: Add coresight nodes")
Change-Id: I87aeff9820628a87a02fac1db94df0b43f5687fd
Signed-off-by: Jie Gan <jie.gan@oss.qualcomm.com>