PENDING: driver: bluetooth: hci_qca: Disable hci_auto_off when BT_EN is consistently pulled up by hardware
If the BT SoC BT_EN is controlled by hardware, disable the AUTO_OFF
feature. Otherwise, BT will close the HCI layer except for the UART
after firmware download. However, the SoC remains active. If the SoC
sends a packet to the Host after firmware download, the Host cannot
respond since the HCI layer is closed, which will cause the firmware
to enter an incorrect state.


Change-Id: If40b580bd9e87153b16c4165181e44a9dd6049c9
Signed-off-by: Shuai Zhang <quic_shuaz@quicinc.com>