PENDING: arm64: dts: qcom: sa7255p: Add address-cells and size-cells property to interrupt-controller node
Updated the missing address-cells and size-cells property in the GIC v3
interrupt controller. This ensures proper parsing of interrupts for
child nodes.
During the parsing of the interrupt map for nodes like pcie, interrupt
parent node are checked for the address-cells property. If the interrupt
controller node does not have the address-cells property defined, the
value will be considered as 0, leading to incorrect parsing of the
interrupt map.

Change-Id: I6eb4bcb8d3a50af9a77a5b2f27bf3249da0fe606
Signed-off-by: Jai Kumar Gautam <quic_jgautam@quicinc.com>