QCLINUX: usb: dwc3: Fix reset assertion during probe failure
The current design sees through that the resets are handled by glue
but in event of probe failure, leaving the reset asserted can lead
to gdsc not turning on before the next probe attempt and hence no
further attempts to probe device will be done. Fix this by leaving
reset de-asserted in case of module removal or porbe deferral.

Change-Id: I8a2616cc7c99eabc77d88a4e604d7b70cfadd4c5
Signed-off-by: Krishna Kurapati <krishna.kurapati@oss.qualcomm.com>