PENDING: misc: fastrpc: prevent context use-after-free in poll mode
In poll mode, the fastrpc driver polls on a memory location that the
DSP updates upon task completion. Once the poll completes, the driver
frees the associated context.

However, there is a race condition where the DSP may send a message
around the same time, which is received in fastrpc_rpmsg_callback().
This callback attempts to access the same context that may have just
been freed by the polling thread.

This race can lead to a use-after-free condition. Defer context
release until after it is guaranteed that the callback will not
access it.

Change-Id: Ia7a4498aef70c6c808a4e9667d25525e75d3d8fe
Signed-off-by: Ling Xu <quic_lxu5@quicinc.com>