arm64: dts: rubikpi3: Adjust the power-on time of the M.2
Some PCIe devices (such as AI accelerator cards) were not enumerated
because insufficient initialization time was given to the devices.

Previously, the power was turned on during RC initialization,
and at least 50 ms of preparation time was reserved for the devices.
Now, the preparation time is increased, and the power is turned on
during regulator-related initialization.

IssueID: RUBIKPI-1082

Change-Id: Id4e2d93038e14874be9d516a0ce51a478b526e44
Signed-off-by: hongyang.zhao <hongyang.zhao@thundersoft.com>